+++ code/core/src/core-events/data/testing-module.ts
@@ -8,22 +8,11 @@ export type TestProviderState = Addon_TestProviderState;
 
 export type TestProviders = Record<TestProviderId, TestProviderConfig & TestProviderState>;
 
-export type TestingModuleRunRequestStories = {
-  id: string;
-  name: string;
-};
-
 export type TestingModuleRunRequestPayload = {
   providerId: TestProviderId;
-  payload: {
-    stories: TestingModuleRunRequestStories[];
-    importPath: string;
-    componentPath: string;
-  }[];
-};
-
-export type TestingModuleRunAllRequestPayload = {
-  providerId: TestProviderId;
+  // TODO: Avoid needing to do a fetch request server-side to retrieve the index
+  indexUrl: string; // e.g. http://localhost:6006/index.json
+  storyIds?: string[]; // ['button--primary', 'button--secondary']
 };
 
 export type TestingModuleProgressReportPayload =