+++ code/core/src/types/modules/addons.ts
@@ -1,11 +1,12 @@
 /* eslint-disable @typescript-eslint/naming-convention */
 import type { FC, PropsWithChildren, ReactElement, ReactNode } from 'react';
 
-import type { TestingModuleProgressReportProgress } from '../../core-events';
+import type { ListItem } from '../../components';
+import type { TestProviderConfig, TestingModuleProgressReportProgress } from '../../core-events';
 import type { RenderData as RouterData } from '../../router/types';
 import type { ThemeVars } from '../../theming/types';
 import type { API_SidebarOptions } from './api';
-import type { API_StatusState, API_StatusUpdate } from './api-stories';
+import type { API_HashEntry, API_StatusState, API_StatusUpdate } from './api-stories';
 import type {
   Args,
   ArgsStoryFn as ArgsStoryFnForFramework,
@@ -28,6 +29,7 @@ export type Addon_Types = Exclude<
   Addon_TypesEnum,
   | Addon_TypesEnum.experimental_PAGE
   | Addon_TypesEnum.experimental_SIDEBAR_BOTTOM
+  | Addon_TypesEnum.experimental_TEST_PROVIDER
   | Addon_TypesEnum.experimental_SIDEBAR_TOP
 >;
 
@@ -329,7 +331,7 @@ export type Addon_Type =
   | Addon_WrapperType
   | Addon_SidebarBottomType
   | Addon_SidebarTopType
-  | Addon_TestProviderType;
+  | Addon_TestProviderType<Addon_TestProviderState>;
 export interface Addon_BaseType {
   /**
    * The title of the addon. This can be a simple string, but it can also be a
@@ -472,8 +474,15 @@ export interface Addon_TestProviderType<
   /** The unique id of the test provider. */
   id: string;
   name: string;
-  title: (state: Addon_TestProviderState<Details>) => ReactNode;
-  description: (state: Addon_TestProviderState<Details>) => ReactNode;
+  /** @deprecated Use render instead */
+  title?: (state: TestProviderConfig & Addon_TestProviderState<Details>) => ReactNode;
+  /** @deprecated Use render instead */
+  description?: (state: TestProviderConfig & Addon_TestProviderState<Details>) => ReactNode;
+  render?: (state: TestProviderConfig & Addon_TestProviderState<Details>) => ReactNode;
+  sidebarContextMenu?: (options: {
+    context: API_HashEntry;
+    state: Addon_TestProviderState<Details>;
+  }) => ReactNode;
   mapStatusUpdate?: (
     state: Addon_TestProviderState<Details>
   ) => API_StatusUpdate | ((state: API_StatusState) => API_StatusUpdate);
@@ -511,7 +520,7 @@ export interface Addon_TypesMapping extends Record<Addon_TypeBaseNames, Addon_Ba
   [Addon_TypesEnum.experimental_PAGE]: Addon_PageType;
   [Addon_TypesEnum.experimental_SIDEBAR_BOTTOM]: Addon_SidebarBottomType;
   [Addon_TypesEnum.experimental_SIDEBAR_TOP]: Addon_SidebarTopType;
-  [Addon_TypesEnum.experimental_TEST_PROVIDER]: Addon_TestProviderType;
+  [Addon_TypesEnum.experimental_TEST_PROVIDER]: Addon_TestProviderType<Addon_TestProviderState>;
 }
 
 export type Addon_Loader<API> = (api: API) => void;