+++ code/frameworks/experimental-nextjs-vite/template/stories/Navigation.stories.tsx
@@ -109,7 +109,7 @@ export default {
   },
 } as Meta<typeof Component>;
 
-export const Default: StoryObj<typeof Component> = {
+export const Default: Story = {
   play: async ({ canvasElement, step }) => {
     const canvas = within(canvasElement);
     const routerMock = getRouter();