+++ code/frameworks/react-native-web-vite/src/vite-plugin.ts
@@ -1,0 +1,16 @@
+import react from '@vitejs/plugin-react';
+
+import { reactNativeWeb } from './preset';
+
+export const storybookReactNativeWeb = () => {
+  return [
+    react({
+      babel: {
+        babelrc: false,
+        configFile: false,
+      },
+      jsxRuntime: 'automatic',
+    }),
+    reactNativeWeb(),
+  ];
+};