+++ test-storybooks/portable-stories-kitchen-sink/react/.storybook/main.ts
@@ -3,8 +3,8 @@ import type { StorybookConfig } from "@storybook/react-vite";
 const config: StorybookConfig = {
   stories: ["../stories/**/*.stories.@(js|jsx|mjs|ts|tsx)"],
   addons: [
+    "@storybook/addon-controls",
     "@storybook/experimental-addon-test",
-    "@storybook/addon-controls"
   ],
   framework: {
     name: "@storybook/react-vite",